feature(133322): Transfere os estilos para estilos compartilhados
